grabbed some of Greg's suggestions and added some others
